AI_Agent/requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==0.0.8
langchain-anthropic==0.1.15  # Specific integration for Anthropic models
pydantic==2.12.4
python-dotenv
wikipedia
duckduckgo-search